refactor: shorten on-chain TLC reconcile module name
