+++ crates/fiber-lib/src/fiber/mod.rs
@@ -13,7 +13,7 @@ mod fee;
 pub(crate) mod gossip_policy;
 mod in_flight_ckb_tx_actor;
 mod key;
-mod onchain_tlc_reconciliation;
+mod onchain_tlc_reconcile;
 #[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod onion_service;
 mod path;
